user location

+++ app/src/main/java/wpics/alcogait/data/DrinksDao.kt
@@ -7,10 +7,10 @@ import androidx.room3.Query
 @Dao
 interface DrinksDao {
     @Query("SELECT * FROM drinks WHERE user_id = :userId")
-    suspend fun getDrinksByUserId(userId: Int): List<Drinks>
+    suspend fun getDrinksByUserId(userId: Long): List<Drinks>
 
     @Query("SELECT * FROM drinks WHERE user_id = :userId AND timestamp = :timestamp")
-    suspend fun getDrinksByUserIdAndTimestamp(userId: Int, timestamp: String): Drinks?
+    suspend fun getDrinksByUserIdAndTimestamp(userId: Long, timestamp: String): Drinks?
 
     @Insert
     suspend fun insertDrink(drink: Drinks)
